RDKEVD-924: return HDMI_CEC_IO_OPERATION_NOT_SUPPORTED when appropriate
Reason for change: To match other halifs
HDMI_CEC_IO_OPERATION_NOT_SUPPORTED should be returned from a function
if the function is not supported on a source device and the function is
being called on a source device.
Test Procedure: Check that halif documentation is as expected.
Priority: P2
Risks: Low